requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
mysql-connector-python==8.3.0
python-dotenv==1.0.1
pydantic==2.6.4
torch==2.2.2
torchvision==0.17.2
transformers==4.46.1
huggingface_hub==0.25.2
diffusers==0.27.2
accelerate==0.29.3
moviepy==1.0.3
edge-tts==6.1.9
requests==2.31.0
Pillow==10.3.0
numpy==1.26.4
google-genai
scipy>=1.12.0
imageio-ffmpeg>=0.4.9
sentencepiece>=0.2.0
mutagen>=1.47.0
parler-tts
